+++ lib/widget/material.dart
@@ -236,4 +236,29 @@ extension WidgetMaterialEx on Widget {
         enableFeedback: enableFeedback,
         child: this,
       );
+
+  Card inCard({
+    Key? key,
+    Color? color,
+    Color? shadowColor,
+    Color? surfaceTintColor,
+    double? elevation,
+    ShapeBorder? shape,
+    bool borderOnForeground = true,
+    EdgeInsetsGeometry? margin,
+    Clip? clip,
+    bool semanticContainer = true,
+  }) =>
+      Card(
+        key: key,
+        color: color,
+        shadowColor: shadowColor,
+        surfaceTintColor: surfaceTintColor,
+        elevation: elevation,
+        shape: shape,
+        borderOnForeground: borderOnForeground,
+        margin: margin,
+        clipBehavior: clip,
+        child: this,
+      );
 }